requirements.txt
pandas==2.2.3
pyarrow>=15,<18         
openpyxl==3.1.3          
xlrd==2.0.1            
boto3==1.35.99          
SQLAlchemy==1.4.54       
psycopg2-binary==2.9.10 
PyYAML==6.0.2           
requests==2.32.3        